src/train/requirements.txt
pymystem3==0.2.0
Telethon==1.19.1
spacy==2.2.3
pandas==1.2.1
matplotlib==3.3.4
numpy==1.20.0
tqdm==4.56.0
seaborn==0.11.1
auto_mix_prep==0.2.0
beautifulsoup4==4.9.3
python-dotenv==0.15.0
PyYAML==5.4.1
